+++ src/renderer/map/style/default-style.js
