+++ src/renderer/map/style/polygon-labels.js
